requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
amqp==5.0.2; python_version >= '3.6'
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0
certifi==2020.12.5
chardet==3.0.4
click-didyoumean==0.0.3
click-repl==0.1.6
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
coreapi==2.3.3
coreschema==0.0.4
xlwt==1.3.0
django-cors-headers==3.4.0
django-rest-swagger==2.2.0
django==3.1
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
docxtpl==0.6.3
drf-yasg2==1.19.4
hashids==1.3.1
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
kombu==5.0.2; python_version >= '3.6'
lxml==4.6.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
openapi-codec==1.3.2
packaging==20.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pillow
prompt-toolkit==3.0.8; python_full_version >= '3.6.1'
psycopg2-binary==2.9.1
pyjwt==1.7.1
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-docx==0.8.7
python-http-client==3.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
redis==3.5.3
requests==2.25.0
sendgrid==6.4.8
simplejson==3.17.2; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uwsgi==2.0.19.1
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
xlrd==1.2.0
mailjet-rest==1.3.4
django-guardian==2.4.0

